# openapi.yaml
openapi: 3.0.3
info:
  title: Customer Intake API Suite
  description: >
    Mock host for Credit Report Application — Stage 1 Customer Intake APIs:
    Customer Details (Siebel CRM), Risk Profile (Risk Engine), Transaction History (Payments Hub).
    All data is fictitious. Requires header X-API-Key.
  version: "1.0.0"
servers:
  - url: /
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Money:
      type: object
      properties:
        value: { type: number, format: double }
        currency: { type: string, example: EUR }
    ErrorResponse:
      type: object
      properties:
        status: { type: integer }
        correlationId: { type: string }
        timestamp: { type: string, format: date-time }
        path: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              title: { type: string }
              detail: { type: string }
              severity: { type: string }
              source: { type: string }
  parameters:
    XCorrelationId:
      name: X-Correlation-ID
      in: header
      schema: { type: string }
      required: false
paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        "200": { description: OK }
  /customer-intake/v1/customers/{id}:
    get:
      summary: API-01 Fetch Customer Details
      description: Retrieve unified party (organisation/person) profile from Siebel CRM.
      parameters:
        - $ref: '#/components/parameters/XCorrelationId'
        - name: id
          in: path
          required: true
          schema: { type: string }
          example: "110023456"
        - name: idType
          in: query
          schema: { type: string, enum: [RABO_CUSTOMER_ID, SIEBEL_ID], default: RABO_CUSTOMER_ID }
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              examples:
                organisation: { value: { customer: { customerId: "110023456", partyType: "ORGANISATION" } } }
        "404": { description: Customer not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        "422": { description: Customer status does not permit intake, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /customer-intake/v1/customers/{customerId}/risk-profile:
    get:
      summary: API-02 Fetch Customer Risk Profile
      description: Retrieve credit risk rating, regulatory classification, AML/KYC risk and bureau data.
      parameters:
        - $ref: '#/components/parameters/XCorrelationId'
        - name: customerId
          in: path
          required: true
          schema: { type: string }
          example: "110023456"
      responses:
        "200": { description: Risk profile found }
        "404": { description: Risk profile not available, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        "423": { description: Access blocked (sanctions match), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /customer-intake/v1/customers/{customerId}/transactions:
    get:
      summary: API-03 Fetch Transaction History
      description: Retrieve booked transaction history and monthly aggregates per account.
      parameters:
        - $ref: '#/components/parameters/XCorrelationId'
        - name: customerId
          in: path
          required: true
          schema: { type: string }
          example: "110023456"
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - name: iban
          in: query
          schema: { type: string }
        - name: bookingStatus
          in: query
          schema: { type: string, enum: [BOOKED, PENDING] }
        - name: pageNumber
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 200 }
      responses:
        "200": { description: Transaction history found }
        "404": { description: No transaction data found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }